final commit

+++ admin/delete.php
@@ -2,7 +2,6 @@
 
     session_start();
 
-
     if (!isset($_SESSION['email'])) {
         header("Location: ../auth/logout.php");
     } elseif ($_SESSION['role'] !== "admin") {
@@ -18,10 +17,39 @@
 
     $deleterecord = $_GET['id'];
 
-    $sql_delete = "DELETE FROM `users` WHERE id = " . $deleterecord;
-    $connect->query($sql_delete);
-    $_SESSION['msg'] = "<div class='alert alert-danger' role='alert' id='alert'>record delete successfully</div>";
-    header("Location: ./dashboard.php");
-    exit();
+    $stmt = $connect->prepare("SELECT * FROM users WHERE id = :id");
+    $stmt->execute([':id' => $deleterecord]);
+    $user = $stmt->fetch(PDO::FETCH_ASSOC);
+
+    if ($user['role'] === "admin") {
+        $_SESSION['toast'] = "<div class='toast show' role='alert' id='toastmsg' aria-live='assertive' aria-atomic='true'>
+                            <div class='toast-header bg-danger'>
+                                <strong class='me-auto'>Notification</strong>
+                                <button type='button' class='btn-close' data-bs-dismiss='toast' aria-label='Close'></button>
+                            </div>
+                            <div class='toast-body bg-danger'>
+                                can not admin user
+                            </div>
+                        </div>";
+        header("Location: ./dashboard.php");
+        exit();
+    } else {
+
+        $sql_delete = "DELETE FROM `users` WHERE id = " . $deleterecord;
+        $connect->query($sql_delete);
+
+        //$_SESSION['msg'] = "<div class='alert alert-danger' role='alert' id='alert'>record delete successfully</div>";
+        $_SESSION['toast'] = "<div class='toast show' role='alert' id='toastmsg' aria-live='assertive' aria-atomic='true'>
+                            <div class='toast-header bg-danger'>
+                                <strong class='me-auto'>Notification</strong>
+                                <button type='button' class='btn-close' data-bs-dismiss='toast' aria-label='Close'></button>
+                            </div>
+                            <div class='toast-body bg-danger'>
+                                user delete successfully
+                            </div>
+                        </div>";
+        header("Location: ./dashboard.php");
+        exit();
+    }
 
 ?>
